+++ src/amr_criteria/amr_criteria.hpp
@@ -10,8 +10,8 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
-#ifndef REFINEMENT_AMR_CRITERIA_HPP_
-#define REFINEMENT_AMR_CRITERIA_HPP_
+#ifndef AMR_CRITERIA_AMR_CRITERIA_HPP_
+#define AMR_CRITERIA_AMR_CRITERIA_HPP_
 
 #include <memory>
 #include <string>
@@ -57,4 +57,4 @@ struct AMRSecondDerivative : public AMRCriteria {
 
 } // namespace parthenon
 
-#endif // REFINEMENT_AMR_CRITERIA_HPP_
+#endif // AMR_CRITERIA_AMR_CRITERIA_HPP_